backend/requirements.txt
Flask
python-dotenv
flask-cors
openai
python-dotenv